+++ examples/theming/component.js
@@ -5,15 +5,21 @@ import EStyleSheet from 'react-native-extended-stylesheet';
 export default class extends React.Component {
   render() {
     return (
-      <View>
-        <Text style={styles.text}>Reload app to see different themes!</Text>
+      <View style={styles.container}>
+        <Text style={styles.text}>Welcome to Extended StyleSheet!</Text>
       </View>
     );
   }
 }
 
 const styles = EStyleSheet.create({
+  container: {
+    flex: 1,
+    justifyContent: 'center',
+    alignItems: 'center',
+    backgroundColor: '$bgColor',
+  },
   text: {
-    color: '$colorPrimary'
+    color: '$textColor',
   }
 });