+++ examples/theming/dark.js
@@ -1,0 +1,5 @@
+export default {
+  $theme: 'dark',
+  $textColor: 'white',
+  $bgColor: 'black',
+}