+++ examples/theming/light.js
@@ -1,0 +1,5 @@
+export default {
+  $theme: 'light',
+  $textColor: 'black',
+  $bgColor: 'white',
+}